use full path when checking deps during compile
This still didn't solve the real problem - we're getting deps out of the
on-the-fly, unpacked package, instead of the packages directory, so the
last write time is always later than the one in targets.

Planning on rewriting to leverage assembly loading hook to load
assemblies from packages instead, or something like that.